refactor: better code arch for future stuff
