+++ argocue/run.go
@@ -5,7 +5,7 @@ import (
 	"github.com/darklab8/argocd-cue/argocue/identifier"
 	"github.com/darklab8/argocd-cue/argocue/logus"
 	"github.com/darklab8/argocd-cue/argocue/manifests"
-	"github.com/darklab8/argocd-cue/argocue/types"
+	"github.com/darklab8/argocd-cue/argocue/settings"
 	"github.com/darklab8/go-typelog/typelog"
 	"github.com/darklab8/go-utils/goutils/utils/utils_types"
 )
@@ -15,7 +15,7 @@ type Deployment interface {
 	GetParameters(utils_types.FilePath)
 }
 
-func Run(workdir utils_types.FilePath, command types.Command) {
+func Run(workdir utils_types.FilePath, command settings.Command) {
 
 	package_type := identifier.IdentifyDeployment(workdir)
 	var deployment Deployment
@@ -30,11 +30,11 @@ func Run(workdir utils_types.FilePath, command types.Command) {
 	}
 
 	switch command {
-	case types.Commands.Generate:
+	case settings.Commands.Generate:
 		deployment.Generate(workdir)
-	case types.Commands.GetParameters:
+	case settings.Commands.GetParameters:
 		deployment.GetParameters(workdir)
 	default:
-		logus.LogStdout.Fatal("not chosen command", typelog.Any("commands", types.Commands))
+		logus.LogStdout.Fatal("not chosen command", typelog.Any("commands", settings.Commands))
 	}
 }
